Temp fix for morphling getting the flash modifier.
Just added morphling to the exclusion check when looping through the list of crew members in order to give them modifiers. When morphling gets flash it is almost impossible to win much like the glitch role. I think some better logic could be employed but I just went through the code quickly to understand it and this seems like a simple addition without much rewriting.